requirements.txt
textual>=0.47.0
rich>=13.0.0
anthropic>=0.25.0
google-generativeai>=0.5.0
openai>=1.0.0
keyring>=24.0.0
aiofiles>=23.0.0
psutil>=5.9.0
click>=8.0.0
pydantic>=2.0.0
requests>=2.31.0
pynvml>=11.0.0
pywebview>=5.0
